+++ pipfreeze/package.py
@@ -67,6 +67,7 @@ class Package(object):
         self._req = Requirement(pinned_requirement)
         self.deprecated = False
         self.deprecation_reason = None
+        self.deprecator = None
 
         # Ensure we're dealing with an exact package version
         if len(self._req.specifiers) != 1 or \
@@ -111,6 +112,7 @@ class Package(object):
     def version(self):
         return self._req.specifiers[0][1]
 
-    def deprecate(self, reason=None):
+    def deprecate(self, reason=None, deprecator=None):
         self.deprecated = True
         self.deprecation_reason = reason
+        self.deprecator = deprecator